+++ Node-Beaver/Node-Beaver.cydsn/sd_manager.h
@@ -9,8 +9,8 @@
 CY_ISR_PROTO(power_interrupt);
 void sd_init();
 //void sd_push(const DataPacket* data_queue, uint8_t data_pos);
-void sd_push(const DataPacket* data_queue, uint8_t data_head,
-             uint8_t data_tail);
+void sd_push(const DataPacket* data_queue, uint16_t data_head,
+             uint16_t data_tail);
 void sd_stop();
